feat(build): make ki use lib.rs
This allows integration tests (a future addition) to use code in ki, it
allows other (future) crates to import code in ki, and it's explicitly
called out as a [pattern in the rust book][1].

[1]: https://doc.rust-lang.org/book/ch11-03-test-organization.html#integration-tests-for-binary-crates
